Update: Functional but needs technical fixes and improvements!
There is possibility for a path lock to happen "No Paths Available" if
the the end of a path is directly next to an occupied cell: if per say
the snake is moving downwards and the last cell of the path's position 
points downwards and directly under the cell there is an occupied cell:
A path creation may not be made due to the fact that either no neighbors
were produced or the algorithm didnt suceed to find a path fast enough.
This leads to the snake colliding with its own body. The other proble
may happen once a new path computation is performed. There is a small
possibility that the path computation is performed just as the the snake
is about to make a critical turn that prevents it from intersecting an
occupied cell. A new path creation is attempted but fails due to similar
circumstances seem in the previous problem.