Reimplementation of PathWrappers complete!
The AI path finding is nearly perfect in the sense of what the goal is.
Currently the main issue is related to updates happening right as a path
is computed. This leads to a small gap which an lead the snake to death
if the head is towards the body with no space in between when the update
happens. Could be solve by adding a check that makes sure that no time
update is made within a specified time frame from when an objective was
reached. When withing this time frame. The updates will wait.!